fix tests

+++ test_bot.py
@@ -12,7 +12,7 @@ class BotTestCase(TestCase):
     bot = PurpleBot()
 
     def setUp(self):
-        PurpleBot._get_rand = lambda self, max: max/2
+        PurpleBot._get_rand = lambda self, max: int(max/2)
 
     def test_roll(self):
         self.assertEqual(10, self.bot._get_roll(20))
